requirements.txt
asgiref==3.7.2
azure-communication-identity==1.3.1
azure-communication-phonenumbers==1.1.0
azure-communication-sms==1.0.1
azure-core==1.29.3
azure-identity==1.14.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
isodate==0.6.1
msal==1.23.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.1
portalocker==2.7.0
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
